% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_qc_metrics.R
\name{calc_simpson_ens}
\alias{calc_simpson_ens}
\title{Calculate Simpson Effective Number of Species (Hill number q=2)}
\usage{
calc_simpson_ens(abund_matrix)
}
\arguments{
\item{abund_matrix}{Numeric matrix of abundances (rows = features, columns = samples).}
}
\value{
Numeric vector of Simpson ENS values, one per sample.
}
\description{
Converts Simpson's index to effective number of species using the inverse
Simpson formula (1/sum(p^2)). This metric is more sensitive to dominant
species than Shannon ENS.
}
\examples{
data(example_feature_table)
result <- calc_simpson_ens(as.matrix(example_feature_table[, -1]))
head(result)
}
