% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_io.R
\name{.convert_output_back_to_original_class}
\alias{.convert_output_back_to_original_class}
\title{Convert filtered table back to original input class}
\usage{
.convert_output_back_to_original_class(
  filtered_table,
  original_class,
  original_object
)
}
\arguments{
\item{filtered_table}{Filtered feature table (data.frame)}

\item{original_class}{Original input class}

\item{original_object}{Original input object (phyloseq or TSE)}
}
\value{
A list with original_table and filtered_table in original class format
}
\description{
Converts the filtered data.frame back to phyloseq or TreeSummarizedExperiment
if the original input was one of those classes.
}
