% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_reporting.R
\name{.generate_markdown_report}
\alias{.generate_markdown_report}
\title{Generate Markdown version of the filtering report}
\usage{
.generate_markdown_report(
  original_stats,
  qc_metrics,
  presence_stats,
  filtering_steps,
  sparsity_elbow_result,
  depth_sparsity_result,
  scree_result,
  output_dir,
  prefix,
  input_description,
  pipeline_params,
  filtered_stats
)
}
\description{
Generate Markdown version of the filtering report
}
