% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_depth_sparsity.R
\name{plot_depth_sparsity}
\alias{plot_depth_sparsity}
\title{Plot depth-sparsity relationship with outlier detection}
\usage{
plot_depth_sparsity(
  analysis_result,
  main = "Depth-Sparsity Relationship",
  color = "blue"
)
}
\arguments{
\item{analysis_result}{Output from \code{\link{analyze_depth_sparsity}}}

\item{main}{Main title for the plot}

\item{color}{Theme color. Default is "blue". Options: "blue", "green", "red", "purple"}
}
\value{
A ggplot object
}
\description{
Creates visualization of the relationship between sequencing depth and
sparsity/richness, with outliers highlighted.
}
\examples{
data(example_feature_table)
analysis <- analyze_depth_sparsity(example_feature_table, verbose = FALSE)
plot_depth_sparsity(analysis)
}
