% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_sparsity_elbow.R
\name{plot_sparsity_elbow}
\alias{plot_sparsity_elbow}
\title{Plot sparsity elbow analysis results}
\usage{
plot_sparsity_elbow(elbow_result, main = "Sparsity Elbow Analysis")
}
\arguments{
\item{elbow_result}{Output from \code{\link{identify_sparsity_elbow}}}

\item{main}{Main title for the plot}
}
\value{
A ggplot object
}
\description{
Creates visualization of the richness-depth relationship with elbow point marked.
}
\examples{
data(example_feature_table)
elbow <- identify_sparsity_elbow(example_feature_table)
plot_sparsity_elbow(elbow)
}
